// src/app/models/Prisoner.ts
export class Prisoner{
    firstname: string;
    lastname: string;
    dateOfBirth: Date;
    nationalID: number;
    education: string;
    streetnumber: number;
    streetname: string;
    suburb: string;
    city: string;
    postalcode: number;
    prison: string;
    location: string;
    offence: string;
    fileNumber: Number;
    status: string;
}